requirements.txt
tqdm
rich
ninja
numpy==1.26.3
pandas
scipy
scikit-learn
scikit-image
matplotlib
opencv-python
imageio
imageio-ffmpeg
open3d
PyOpenGL
glfw

einops
tensorboard
tensorboardX

# for gui
dearpygui

# for grid_tcnn
# git+https://github.com/NVlabs/tiny-cuda-nn/#subdirectory=bindings/torch

# for stable-diffusion
huggingface_hub
diffusers >= 0.9.0
accelerate
transformers

# for dmtet and mesh export
xatlas
trimesh
pyglet
PyMCubes
pymeshlab
git+https://github.com/NVlabs/nvdiffrast/

# for zero123
carvekit-colab
omegaconf
pytorch-lightning==2.1.3
# taming-transformers-rom1504
kornia
git+https://github.com/openai/CLIP.git

# for omnidata
gdown

# for dpt
timm

# for remote debugging
debugpy-run

# for deepfloyd if
sentencepiece

# logging
wandb